+++ NEWS.md
@@ -1,5 +1,6 @@
 # esri2sf (development version)
 
+* Add `{arcgisutils}` to Suggests. `esri2sf()` also now uses `arcgisutils::arc_token()` (if installed) to set the default value of token. Set token to `""` to override this potentially breaking change.
 * Added `esriitem()` and `esiruser()` functions supporting content and community APIs. (2023-10-23)
 * Added `glue_ansi_sql()`, `glue_sql_bbox()`, `fmt_epoch_date()`, and `fmt_epoch_dates()` helper functions. (2023-08-05)
 * Allow input `url` values that contain a trailing query. This query is stripped from the input URL by default. (2023-08-22) Also add `trimws()` to support URLs with trailing or leading white space. (2023-10-23)